+++ pinet-rest/src/main/java/com/pinet/rest/service/ICartComboDishSpecService.java
@@ -28,7 +28,7 @@ public interface ICartComboDishSpecService extends IService<CartComboDishSpec> {
      * 删除购物车套餐明细
      * @param cartId
      */
-     void deleteByCartId(Integer cartId);
+     void deleteByCartId(Long cartId);
 
     /**
      * 批量删除购物车套餐明细
